refactor: 당첨 번호 예외처리 클래스 리팩터링
1. WinNumber 클래스 기능 => Lotto 클래스로 기능 통합
2. WinNumber.test => LottoTest 로 기능 통합
