+++ __tests__/LottoTest.js
@@ -1,19 +1,41 @@
-const Lotto = require("../src/Lotto");
+const Exception = require("../src/error/exception");
+const Lotto = require("../src/error/Lotto");
 const { ERROR } = require("../src/utils/constant");
 
 describe("로또 클래스 테스트", () => {
   test("로또 번호의 개수가 6개가 넘어가면 예외가 발생한다.", () => {
     expect(() => {
-      new Lotto([1, 2, 3, 4, 5, 6, 7]).validate();
+      new Lotto([1, 2, 3, 4, 5, 6, 7]).checkInput();
     }).toThrow(ERROR.DUPLICATE);
   });
 
-  // TODO: 이 테스트가 통과할 수 있게 구현 코드 작성
   test("로또 번호에 중복된 숫자가 있으면 예외가 발생한다.", () => {
     expect(() => {
-      new Lotto([1, 2, 3, 4, 5, 5]).validate();
+      new Lotto([1, 2, 3, 4, 5, 5]).checkInput();
     }).toThrow(ERROR.DUPLICATE);
   });
 
-  // 아래에 추가 테스트 작성 가능
+  test("당첨 번호 문자 길이 예외처리", () => {
+    const input = "1";
+    const exception = new Exception();
+    const errorTest = () => exception.isAllow(new Lotto(input));
+
+    expect(errorTest).toThrow(ERROR.WIN_NUMBER);
+  });
+
+  test("당첨 번호 개수 예외처리", () => {
+    const input = "1,2,3,4,5,6,7";
+    const exception = new Exception();
+    const errorTest = () => exception.isAllow(new Lotto(input));
+
+    expect(errorTest).toThrow(ERROR.WIN_NUMBER);
+  });
+
+  test("당첨 번호 유효 숫자 범위 예외처리", () => {
+    const input = "2,4,8,16,32,64";
+    const exception = new Exception();
+    const errorTest = () => exception.isAllow(new Lotto(input));
+
+    expect(errorTest).toThrow(ERROR.WIN_NUMBER);
+  });
 });
